Generalize iteration over backedges
This prepares the way for allowing a MethodInstance
to store dispatch tuples as well as other MethodInstances
among their backedges. The motivation is to properly
handle invalidation with `invoke`, which allows calling
a less-specific method and thus runs afoul of our standard
mechanisms to detect "outdated" dispatch.

This implements the functionality only on the C side,
and has no visible external consequences.