Support customization of termination signal for timeouts
We've been having a lot of timeouts on CI recently.  Our coredumps might
be helpful in tracking these down, but when we send `SIGTERM` or
`SIGKILL` we don't get coredumps.  This allows customization of which
signal is sent during these timeout messages, to allow for CI to set it
to `SIGSEGV` or similar, to force core dumps for debugging purposes.